requirements.txt
apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
crispy-bootstrap5==0.6
cryptography==3.4.8
cssselect2==0.6.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.7.0
distro-info===1.1build1
Django==4.1.1
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-jazzmin==2.5.0
django-smart-selects==1.6.0
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
future==0.18.2
gunicorn==20.1.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.10
oscrypto==1.3.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF3==1.0.6
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2017.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
systemd-python==234
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.2
tzlocal==4.2
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.0
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
whitenoise==6.2.0
xdg==5
xhtml2pdf==0.2.8
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
